// src/types/starships.ts
export type Starship = {
  name: string;
  model: string;
  manufacturer: string;
  starship_class: string;
  cost_in_credits: string;
  length: string;
  crew: string;
  passengers: string;
  cargo_capacity: string;
  max_atmosphering_speed: string;
  hyperdrive_rating: string;
  MGLT: string;
  consumables: string;
  url: string;
};